use eslint and adjust code
